DATAMONGO-880 - Avoid potential NPE in LazyLoadingInterceptor#intercept.
We now check the result of LazyLoadingInterceptor#ensureResolved() before invoking the given method reflectively. If the result is null we directly return null.
Object methods that are not overridden are now executed against the MethodProxy.
Added test case for resolving a dbref that points to a value that is deleted in between. 

Original pull request: #154.